# Optimize batch import to KV backends