Expose a `quad.Writer` constructor on `QuadStore` to allow backend implementations to optimize quad import.

Also, implement the first set of optimizations for KV backends. Results for a 30K quads file:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkBolt/qs/import-8     3212907       490386        -84.74%
BenchmarkLeveldb/qs/import-8  4199793       231921        -94.48%
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cayleygraph/cayley/803)
<!-- Reviewable:end -->